Changed the type of function output on ft_strlen

+++ libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: adarabi <adarabi@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/21 11:56:57 by adarabi           #+#    #+#             */
-/*   Updated: 2026/05/07 00:28:37 by adarabi          ###   ########.fr       */
+/*   Updated: 2026/05/07 00:45:22 by adarabi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *c)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (c[count] != '\0')
